requirement.txt
Pillow
scipy
matplotlib
torch
torchvision